Add point and base builtin classes, restructuring slightly, add `BasicClassesPlugins`
