+++ src/lib.rs
@@ -42,9 +42,7 @@ impl PluginGroup for TrenchBroomPlugins {
 			.add(CorePlugin(self.0))
 			.add(fgd::FgdPlugin)
 			.add(class::QuakeClassPlugin)
-			.add(class::builtin::BasicClassesPlugin)
-			.add(class::builtin::LightingClassesPlugin(default()))
-			.add(class::builtin::SolidClassesPlugin)
+			.add_group(class::builtin::BasicClassesPlugins)
 			.add(config::ConfigPlugin)
 			.add(qmap::QuakeMapPlugin)
 			.add(geometry::GeometryPlugin)
